+++ locales/fr.json
@@ -2,6 +2,6 @@
     "App Title" : "X Launcher Core QT",
     "Main Button" : "Clicker sur moi",
     "Hello World" : "Salut le monde!",
-    "About SysTray" : "À Propos",
+    "About SysTray" : "À Propos de",
     "Quit SysTray" : "Quitter"
 }
