+++ pkg/windows/docs/code-of-conduct.rtf
