+++ pkg/windows/installer-windows.iss
@@ -1,7 +1,7 @@
 ; Script generated by the Inno Setup Script Wizard.
 ; SEE THE DOCUMENTATION FOR DETAILS ON CREATING INNO SETUP SCRIPT FILES!
 
-#define MyAppName "X Launcher Core"
+#define MyAppName "X Launcher Core Legacy"
 #define MyAppVersion "0.0.0.0.2"
 #define MyAppPublisher "Xgui4 Studio"
 #define MyAppURL "https://www.github.com/xgui4/X-Launcher"
@@ -28,27 +28,27 @@ ArchitecturesInstallIn64BitMode=x64compatible
 ChangesAssociations=yes
 DefaultGroupName={#MyAppName}
 AllowNoIcons=yes
-LicenseFile=..\LICENSE
-InfoBeforeFile=..\README.md
-InfoAfterFile=..\code-of-conduct.rtf
+LicenseFile=..\..\LICENSE
+InfoBeforeFile=..\..\README.md
+InfoAfterFile=docs\code-of-conduct.rtf
 ; Uncomment the following line to run in non administrative install mode (install for current user only).
 ;PrivilegesRequired=lowest
 PrivilegesRequiredOverridesAllowed=dialog
 OutputBaseFilename=x-launcher-core
-SetupIconFile=..\X-Launcher.Desktop\app-icon.ico
+SetupIconFile=..\..\assets\app-icon.ico
 SolidCompression=yes
 WizardStyle=modern
 
 [Languages]
-Name: "english"; MessagesFile: "compiler:Default.isl"
-Name: "french"; MessagesFile: "compiler:Languages\French.isl"
+Name: "english"; MessagesFile: "compiler.english.isl"
+Name: "french"; MessagesFile: "compiler:french.isl"
 
 [Tasks]
 Name: "desktopicon"; Description: "{cm:CreateDesktopIcon}"; GroupDescription: "{cm:AdditionalIcons}"; Flags: unchecked
 
 [Files]
-Source: "..\publish\windows\{#MyAppExeName}"; DestDir: "{app}"; Flags: ignoreversion
-Source: "..\X-Launcher\publish\windows\*"; DestDir: "{app}"; Flags: ignoreversion recursesubdirs createallsubdirs
+Source: "..\..\publish\windows\{#MyAppExeName}"; DestDir: "{app}"; Flags: ignoreversion
+Source: "..\..\X-Launcher\publish\windows\*"; DestDir: "{app}"; Flags: ignoreversion recursesubdirs createallsubdirs
 ; NOTE: Don't use "Flags: ignoreversion" on any shared system files
 
 [Registry]
